requirements.txt
async-timeout==5.0.1
bleak==2.1.1
contourpy==1.3.2
cycler==0.12.1
dbus-fast==3.1.2
fonttools==4.61.1
kiwisolver==1.4.9
matplotlib==3.10.8
numpy==1.26.4
packaging==25.0
pillow==12.1.0
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
six==1.17.0
typing_extensions==4.15.0
